requirements.txt
azure-cognitiveservices-speech==1.42.0
cffi==1.17.1
exceptiongroup==1.2.2
filelock==3.17.0
fsspec==2025.3.0
iniconfig==2.0.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.2
pluggy==1.5.0
pycparser==2.22
pytest==8.3.5
python-dotenv==1.0.1
sounddevice==0.5.1
soundfile==0.13.1
sympy==1.13.1
tomli==2.2.1
torch==2.6.0
triton==3.2.0
typing_extensions==4.12.2
